FEATURE: Add server timing header
This header will create additional
output in the browser dev tools.
Each timing will be added as entry
in the timings tab and helps devs
to understand which part of the app
took how long.

Initially it will provide the duration of the
request process, fusion rendering time
and sql execution time.

But via the DebugService additional
timings can be added by the dev.